move upper level info to new upper level folder
