--- tlidb/TLiDB/__init__.py
@@ -1,8 +1,0 @@
-from .version import __version__
-
-supported_datasets = [
-    "clinc150",
-    "DailyDialog",
-    "Friends"
-]
-
