+++ tlidb/__init__.py
@@ -1,0 +1,7 @@
+from .version import __version__
+
+supported_datasets = [
+    "DailyDialog",
+    "Friends",
+    "clinc150",
+]
